Added funcionality to Get all products. Safeguards are in place if there is no data to return.
Added safeguards to functionality to Get product by id.

Added functionality to Add a new product to the database. Safeguards for the possibility of encountering errors in case of repeating primary keys are in place but do work due to internal handling of this kind of errors by EntityFramework. (we cannot catch them so swagger shows code 500 internal server error instead of what we want)

Added data anotations to product model to make id primary key and to make it autogenerated.